+++ packages/backend-airdos/src/index.ts
@@ -1,53 +1,16 @@
-import express from 'express'
-import _ from 'lodash/fp'
-import { ApolloServer, makeExecutableSchema } from 'apollo-server-express'
+import { server } from './lambda'
 
-import { typeDefs, resolvers } from './modules/schema'
-import { getUserModels } from './modules/types/models'
-import { getUserFromToken } from './utils'
-
-require('dotenv').config()
-
-const PORT = 4000
-
-const schema = makeExecutableSchema({
-  typeDefs,
-  resolvers,
-  resolverValidationOptions: { requireResolversForResolveType: false },
-})
-
-const server = new ApolloServer({
-  schema,
-  playground: {
-    endpoint: `http://localhost:${PORT}/graphql`,
-  },
-  context: async ({ req }) => {
-    const token = _.get('headers.token', req)
-    let user = null
-    try {
-      if (token) {
-        user = await getUserFromToken(token)
-      }
-    } catch (e) {
-      console.warn(`Unable to authenticate using token: ${token}`)
-    }
-    return {
-      user,
-      models: {
-        user: getUserModels({ user }),
-      },
-    }
+const graphqlHandler = server.createHandler({
+  cors: {
+    origin: 'https://airdos.net/',
+    credentials: false,
+    allowedHeaders: [
+      'Content-Type',
+      'Origin',
+      'Accept',
+      'token',
+    ],
   },
 })
 
-const app = express()
-
-server.applyMiddleware({
-  app,
-})
-
-// Start the server
-app.listen(PORT, () => {
-  console.log(`The server has started on port: ${PORT}`)
-  console.log(`http://localhost:${PORT}/graphql`)
-})
+export { graphqlHandler }
