+++ packages/backend-airdos/src/modules/enums/PostTypeEnum.ts
@@ -1,4 +1,4 @@
-import { gql } from 'apollo-server-express'
+import { gql } from 'apollo-server-lambda'
 
 export const typeDefs = gql`
   enum PostTypeEnum {
